backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
djoser
psycopg2-binary
python-decouple
django-cors-headers
